Make InMemoryEntityLookup production code
Fixes #29
